requirements.txt
Django==2.2.5
djangorestframework==3.10.3
httpie==1.0.3
mysqlclient==1.4.4
python-decouple==3.1
